#38 feat: Smart Choice API 구조 개선 및 명칭 변경
