+++ src/app/lib/axiosInstance.ts
@@ -1,7 +1,10 @@
 import axios from "axios";
 
-// Axios 인스턴스 생성, 필요 시 환경 변수 기반 기본 설정을 추가할 수도 있음
-const axiosInstance = axios.create({
+/**
+ *  기본 클라이언트 인스턴스
+ *  인증이 필요하지 않은 요청에 사용됨.
+ */
+export const client = axios.create({
   baseURL: "/api", // 내부 API 라우트에 접근
   timeout: 30000,
   headers: { "Content-Type": "application/json", Accept: "application/json" },
@@ -9,7 +12,7 @@ const axiosInstance = axios.create({
 });
 
 // 모든 API 요청 오류를 공통으로 처리하기 위한 전역 interceptor 활용
-axiosInstance.interceptors.response.use(
+client.interceptors.response.use(
   (response) => response,
   (error) => {
     if (axios.isAxiosError(error)) {
@@ -20,5 +23,3 @@ axiosInstance.interceptors.response.use(
     return Promise.reject(error);
   }
 );
-
-export default axiosInstance;
